Let a staged chunk hold a string column larger than 2 GiB
duckdb_chunk_staging accumulated int32 offsets and threw 'string column exceeds cudf int32 offset
limit', which capped a .hpln chunk by the widest string column in the table -- for TPC-H, by
l_comment and ps_comment, which no query reads. The limit was ours, not cudf's: this repo is on
libcudf 26.06, where large strings with 64-bit offsets have been default since ~24.12.

Staging now keeps int64 offsets and builds a 32-bit offsets child while the chars fit in one and a
64-bit child when they do not, so an ordinary column is byte-identical to what it was and an
oversized one is written rather than refused. A 4 GB partsupp chunk now writes 27.1M rows with
ps_comment at ~3.3 G chars, and reading that column back is byte-identical to parquet.

Also logs what a pin chunk actually costs (compressed against decoded bytes), because 'does one
format pin more than the other' was otherwise only answerable by inference from bytes read -- which
is the file's size for a .hpln and the source's for parquet, and neither is the pinned footprint.
It answers the question: for lineitem's 14 read columns a .hpln pin holds 99.02 GB against a
parquet pin's 143.48 GB.

The lift buys no measurable time (lineitem chunks 97 -> 78 against parquet's 73; host-pinned 8.739
-> 8.721 s, cold 49.595 -> 49.873) -- it is kept for the constraint it removes, not for speed.

+++ src/include/helper/duckdb_chunk_staging.hpp
@@ -98,8 +98,11 @@ class duckdb_chunk_staging {
   /// the sources feeding this are either small (VALUES) or bounded by one chunk, so neither a
   /// pinned allocation nor the IO machinery is warranted.
   struct column_staging {
-    std::vector<std::uint8_t> fixed_data;        // fixed-width payload
-    std::vector<std::int32_t> offsets;           // varchar: num_rows + 1 entries
+    std::vector<std::uint8_t> fixed_data;  // fixed-width payload
+    /// varchar: num_rows + 1 entries. int64 because a chunk's chars may exceed INT32_MAX; the
+    /// column is built with a 32-bit offsets child when it fits and a 64-bit one when it does not
+    /// (libcudf large strings), so the common case is byte-identical to what it always was.
+    std::vector<std::int64_t> offsets;
     std::vector<char> chars;                     // varchar payload
     std::vector<cudf::bitmask_type> mask_words;  // cudf validity bitmask (1 = valid)
     cudf::size_type null_count = 0;
